Moving auto inget user preferences out of autopsy core
